Integrate with CloudAMQP and EasyCCG parser
The chatbot now connects to the EasyCCG parser. For each message
addressed at the chatbot, the chatbot extracts the payload (which is
suppposed to be a sentence to parse), sends it to the EasyCCG parser,
waits for the response, and sends the response as a message to the
channel.

For now, the payload is the content of the message after the first ":".

At this stage, no arguments for the parser are allowed. This will be
changed at later stages, in particular, the number of parse trees to
return should be specifiable in the initial message to the chatbot.

Also, at later stages the chatbot should return the parse trees in
default format in the message body, and in HTML format as an attached
HTML file.